chore(extensionKind): set extensionKind to workspace
Setting the extensionKind to workspace allows the extension to be
activated in the context of a workspace. This is useful for users
who uses the extension connecting the VS Code UI to a remote VS Code
server. This way the pylsp used will be always the one installed from
the workspace environment.

Also bump the vscode packages.

Signed-off-by: Matheus Castello <matheus@castello.eng.br>