chore(pylsp): Use pylsp directly as the language server
Previously, the extension was using a TCP connection, but was not
clear how the pylsp was being started. Also the pylsp.executable
was being set, but it's value was not being used.

This patch changes the way the language server is started, now
using the pylsp executable directly, and also uses the pylsp.executable
value to set the command to be executed.

Signed-off-by: Matheus Castello <matheus@castello.eng.br>